Added...so much
In the spirit of bad practice I made a ton of changes in one save
haha... Added a form for admin override if a student is signed in twice.
Made all the fields and buttons work on the main form and new form, and
added 2 more spreadsheets to the database, and made it so that it tracks
sign in numbers.